+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 ## 0.13.0
 
+### New Features
+- New location import plugin
+  * [Little Navmap](https://albar965.github.io/littlenavmap.html) user points CSV import
+
 ### Bug Fixes
 - The various tabs in the Flight information dialog (Description, Aircraft, Conditions and Flight Plan) are now properly updated when importing a flight or loading a flight from the logbook
 
 ### Under The Hood
+- Return value optimisation
 - Rule of Zero, copy-and-swap [[Back to Basics: RAII and the Rule of Zero - Arthur O'Dwyer - CppCon 2019](https://www.youtube.com/watch?v=7Qgd9B1KuMQ)]
 - Thread-safe access to singletons [[Back to Basics: Concurrency - Arthur O'Dwyer - CppCon 2020](https://www.youtube.com/watch?v=F6Ipn7gCOsY)]
 - "Const pimpl" pattern (where applicable) [[CppCon 2016: Herb Sutter “Leak-Freedom in C++... By Default.”](https://www.youtube.com/watch?v=JfmTagWcqoE)]
@@ -15,7 +20,6 @@
 ## 0.12.0
 
 ### New Features
-
 - Location module
   * Capture and teleport to locations in the flight simulator
   * Manually add and edit locations